switch api signup, login to use jwt (now they actually function)
