+++ src/app/api/auth/signup/route.ts
@@ -1,0 +1,63 @@
+import { createAccount } from "@/lib/accountUtils";
+import prisma from "@/lib/db";
+import { hash, verify } from "argon2";
+import jwt from 'jsonwebtoken'
+import { JWT_SECRET } from "@/app/api/config";
+
+
+export async function POST(req: Request) {
+  const { username, password, provider } = await req.json()
+
+  if (!username || !password) {
+    return new Response(JSON.stringify({ error: 'Missing required fields' }), { status: 400 });
+  }
+
+  // Check if the user already exists
+  const existingUser = await prisma.user.findUnique({ where: { name: username } })
+  if (existingUser) {
+    return new Response(JSON.stringify({ error: 'User already exists' }), { status: 400 });
+  }
+
+  const hashedPassword = await hash(password)
+
+  const role = await prisma.role.findFirst({
+    where: { name: 'USER' }
+  })
+
+  // if (!role) {
+  //   return res.status(500).json({ error: 'Default role not found' })
+  // }
+
+  try {
+    const newUser = await prisma.user.create({
+      data: {
+        name: username,
+        username,
+        hashedPassword,
+        roles: {
+          connect: { id: role!.id }
+        }
+      }
+    })
+    const accountData = 
+      provider === 'credentials' ? 
+        {username, hashedPassword}:
+        { accessToken: 'actionTokenReel', refreshToken: 'refreshTokenReel', providerAccountId: 'someOAuthId' } 
+    
+    await createAccount(newUser.id, provider, 'someProviderId', accountData)
+  
+    // return new Response(JSON.stringify({
+    //     id: newUser.id,
+    //     email: newUser.email,
+    //     name: newUser.name,
+    //     createdAt: newUser.createdAt,
+    // }), { status: 201 });
+    const token = jwt.sign({
+      id: newUser.id,
+      name: newUser.name
+    }, JWT_SECRET!, {expiresIn: '7d'})
+    return new Response(JSON.stringify({ token }), {status: 200})
+  } catch (error) {
+    return new Response(error as string, {status: 500})
+  }
+}
