+++ src/app/api/config.ts
@@ -1,0 +1,3 @@
+
+
+export const JWT_SECRET = process.env.NEXTAUTH_SECRET
